move all marker's logic to context
